Unify memory storage APIs
