+++ tests/test_memory_lattice_retrieve.py
@@ -7,7 +7,7 @@ sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 from pro_metrics import tokenize, lowercase  # noqa: E402
 import pro_memory  # noqa: E402
 import pro_rag  # noqa: E402
-from memory.memory_lattice import MemoryGraphStore  # noqa: E402
+from memory import MemoryStore  # noqa: E402
 import pro_predict  # noqa: E402
 
 
@@ -21,7 +21,7 @@ async def test_retrieve_uses_graph_and_embedding(monkeypatch):
     monkeypatch.setattr(
         pro_predict, "_VECTORS", {"dogs": {"d": 1.0}, "cats": {"c": 1.0}}
     )
-    store = MemoryGraphStore()
+    store = MemoryStore()
     store.add_utterance("d1", "user", "cats eat fish", _embed("cats eat fish"))
     store.add_utterance(
         "d1", "user", "dogs chase cats", _embed("dogs chase cats")
